#313 moved IndexedButton to buttons, one minor fix in WithCollectionValueComponentProviderMixin
